Improve edge case handling in UI; Add error boundary for amortization card; Improve loading of devices
