+++ apps/web/src/components/recommendations/amortization/ammortization-card-error.tsx
@@ -1,0 +1,17 @@
+"use client";
+
+import { Card, CardContent, CardHeader, CardTitle } from "@energyleaf/ui/card";
+import { TryAgainErrorHint } from "@energyleaf/ui/error";
+
+export default function AmortizationCardError() {
+    return (
+        <Card>
+            <CardHeader>
+                <CardTitle>Geräte-Amortisation</CardTitle>
+            </CardHeader>
+            <CardContent>
+                <TryAgainErrorHint />
+            </CardContent>
+        </Card>
+    );
+}
